+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Add new `reportUnhandledPromiseRejectionsAsHandled` config option
+
 ## [8.0.0] - 2024-08-29
 
 ### Summary